Modify SMB

+++ KungFu/Assets/Scripts/Animation/BaseSmb.cs
@@ -9,8 +9,8 @@ public class BaseSmb : StateMachineBehaviour
         PlayerAnimController _animCtrl = animator.gameObject.GetComponent<PlayerAnimController>();
         if (_animCtrl != null)   // it is not a enemy
         {
-            animator.SetFloat("AttackID_i", -1);
-            animator.SetFloat("DefenseID_i", -1);
+            animator.SetInteger("AttackID_i", -1);
+            animator.SetInteger("DefenseID_i", -1);
         }
     }
     override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
